uk_ch_intel/requirements.txt
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
requests==2.31.0
httpx==0.25.2
pandas==2.1.4
polars==0.19.19
redis==5.0.1
rq==1.15.1
python-dotenv==1.0.0
tqdm==4.66.1
tenacity==8.2.3
click==8.1.7
pydantic==2.5.3
pydantic-settings==2.1.0
loguru==0.7.2
